+++ CHANGELOG.md
@@ -1,3 +1,20 @@
 ## 0.0.1
 
-* TODO: Describe initial release.
+*  Initial release.
+
+## 0.1.2
+
+* feat: Add onDragEnd optional argument, it is function with argument (Offset) which is current position (related to parent) of top left corner of note | Example usage: storing notes last position to create it later with same position
+* feat: Add optional constructor parameter (Offset) 'startingPosition' which is starting position (related to parent) of top left corner of note
+
+## 0.2.0
+
+* docs: Add documentation to package.
+
+## 0.3.0
+
+* feat!: Add position property which shows current position (Offset) of the note (top left corner of parent is (0,0))
+
+## 0.4.0
+
+* test!: Add unit and widget tests to package